src/main/java/edu/mondragon/aiqua_server/models/UserType.java
package edu.mondragon.aiqua_server.models;

import java.util.List;

import jakarta.persistence.CascadeType;
import jakarta.persistence.Entity;
import jakarta.persistence.Id;
import jakarta.persistence.OneToMany;
import jakarta.persistence.Table;

@Entity
@Table(name = "userType")
public class UserType {
    @Id
    private Integer userTypeID;

    private String description;

    @OneToMany(mappedBy = "userType", cascade = CascadeType.ALL, orphanRemoval = true)
    private List<User> users;

    public UserType() {
    }

    public UserType(Integer userTypeID, String description) {
        this.userTypeID = userTypeID;
        this.description = description;
    }

    public Integer getUserTypeID() {
        return userTypeID;
    }

    public void setUserTypeID(Integer userTypeID) {
        this.userTypeID = userTypeID;
    }

    public String getDescription() {
        return description;
    }

    public void setDescription(String description) {
        this.description = description;
    }
}
